dags/requirements.txt
pandas==1.3.5
numpy==1.21.6
pandera==0.18.3
sqlalchemy==1.4.46
psycopg2-binary==2.9.5